+++ lib/rewrite/source_error.ex
@@ -1,3 +1,0 @@
-defmodule Rewrite.SourceError do
-  defexception [:message]
-end