+++ test/rewrite/source_test.exs
@@ -2,7 +2,6 @@ defmodule Rewrite.SourceTest do
   use ExUnit.Case
 
   alias Rewrite.Source
-  alias Rewrite.SourceError
   alias Sourceror.Zipper
 
   doctest Rewrite.Source
@@ -240,41 +239,6 @@ defmodule Rewrite.SourceTest do
       assert Source.modules(source, 1) == [MyApp.Simple]
       assert Source.modules(source, 2) == [TheApp.Simple]
       assert Source.modules(source, 3) == [AnApp.Simple]
-    end
-  end
-
-  describe "debug_info/2" do
-    test "returns debug info" do
-      source = Source.read!("lib/rewrite/source.ex")
-      dbgi = Source.debug_info(source, Rewrite.Source)
-
-      check =
-        case dbgi do
-          {:ok, _dbgi} -> true
-          {:error, :cover_compiled} -> true
-        end
-
-      assert check
-    end
-
-    test "returns an error tuple for an unknown module" do
-      assert "a + b" |> Source.from_string() |> Source.debug_info(Unknown) ==
-               {:error, :non_existing}
-    end
-  end
-
-  describe "debug_info!/2" do
-    test "returns debug info" do
-      source = Source.read!("test/fixtures/source/simple.ex")
-      assert Source.debug_info!(source, MyApp.Simple)
-    end
-
-    test "raises an error" do
-      source = Source.read!("test/fixtures/source/simple.ex")
-
-      assert_raise SourceError, "Can not find debug info, reason: :non_existing", fn ->
-        Source.debug_info!(source, MyApp.Unknown)
-      end
     end
   end