+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 + Remove `Rewrite.Issue`. The type of the field `issues` for `Rewrite.Source`
   becomes `[term()]`.
++ Remove `Rewrite.Source.debug_info/2` and `BeamFile` dependency.
++ Add `Rewrite.Project.sources_by_module/2`, `Rewrite.Project.source_by_module/2`
+  and `RewriteProject.source_by_module!/2`.
 
 ## 0.1.1 - 2022/09/07